Recreate the Handler hierarchy
Created a GroupMeService in order to properly mock sendMessage.
BaseHandler moved to an interface called Handler.
