+++ api/loki.ts
@@ -3,12 +3,12 @@ import { NowRequest, NowResponse } from "@now/node";
 import { container } from "tsyringe";
 
 import HandlerRegistry from "../src/HandlerRegistry";
-import { GroupMeInfo } from "../src/handlers/BaseHandler";
+import { GroupMeInfo } from "../src/services/GroupMeService";
 
 export default async (req: NowRequest, res: NowResponse): Promise<void> => {
   const groupMeInfo = req.body as GroupMeInfo;
-  const handlerRegistry = container.resolve(HandlerRegistry);
 
+  const handlerRegistry = container.resolve(HandlerRegistry);
   const actions = handlerRegistry.delegate(groupMeInfo);
 
   await Promise.all(actions);
