+++ src/handlers/HeyHandler.ts
@@ -1,24 +1,19 @@
-import BaseHandler, { GroupMeInfo } from "./BaseHandler";
-import HandlerConfig from "./HandlerConfig";
 import { singleton, inject } from "tsyringe";
-import HandlerRegistry from "../HandlerRegistry";
+
+import Handler from "./Handler";
+import HandlerConfig from "./HandlerConfig";
+import { GroupMeService, GroupMeInfo } from "../services/GroupMeService";
 
 @singleton()
-class HeyHandler extends BaseHandler {
+export default class HeyHandler implements Handler {
   private static readonly regexps = [new RegExp("hey loki", "i")];
   readonly config: HandlerConfig = {
     regexps: HeyHandler.regexps,
   };
 
-  public constructor(@inject(HandlerRegistry) handlerRegistry: HandlerRegistry) {
-    super();
-
-    handlerRegistry.register(this);
-  }
+  constructor(@inject(GroupMeService) private readonly groupMeService: GroupMeService) {}
 
   async handle(groupMeInfo: GroupMeInfo): Promise<number> {
-    return this.sendMessage(`Howdy ${groupMeInfo.name}!`);
+    return this.groupMeService.sendMessage(`Howdy ${groupMeInfo.name}!`);
   }
 }
-
-export default HeyHandler;
