+++ tsconfig.json
@@ -17,6 +17,6 @@
   },
   "include": [
     "api/**/*.ts",
-		"src/**/*.ts"
+    "src/**/*.ts"
   ]
 }
